Moved telegraf utils to a separate package under Common for efficient code share
