+++ Common/telegraf_utils/telegraf_name_map.py
